// src/component/model/loanListModel.ts
export class LoanListModel {
  "ApplicationId": number;
  "LoanProductCode": string;
  "LoanProductName": string;
  "AppliedLoanAmount": number;
  "PurposeOfLoan": string;
  "InterestRate": number;
  "TotalInstallment": number;
  "LoanApplicationDate": string;
  "LoanStatus": string;
  "PresentAddress": string;
  "ParmanentAddress": string;
  "LoanNumber": string;
  "ProductTypeId": number;
  "CurrentStageName": string;
  "BranchName": string;
  "FullName": string;
  "PersonId": number;
  "EntryBranchName": string;
  "BranchCode": string;
}